Finished Payroll v3
Added the ability for the user to get Total Pay Owed from the payTotal list. Added the option for the user to print a list of the employees entered into the namesList. V4 will include persistant data for multiple sessions.

Also cleaned up the other versions where i experimented with various features that made it into v3 final.